Server Controller Function segregation
